requirements.txt
# Essential build tools (install first)
setuptools>=65.0.0
wheel
pip>=22.0

# Core framework
streamlit>=1.28.0

# LangChain ecosystem (using more recent stable versions)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Document processing
unstructured[local-inference]>=0.10.0
pypdf>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web scraping and requests
requests>=2.31.0

# Environment and utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# Additional dependencies
numpy>=1.24.0
tiktoken>=0.5.0

# HTTP client
httpx>=0.25.0